Added overloaded new and delete for stack allocation -- use new (stackalloc) typename to create an object through _malloca. Added the stack_ptr class, which is basically an auto_ptr for pointers allocated via stackalloc new. Switched all configurations to /W4. Added InternalHelpers.h with macros to help us build cleanly in /W4.

+++ trunk/source/directsound/WaveFormat.cpp
@@ -95,7 +95,7 @@ namespace DirectSound
 			m_ExtraData = gcnew DataStream( data + offset, extraDataBytes, true, true, false );
 		}
 
-		m_Format->cbSize = extraSize + extraDataBytes;
+		m_Format->cbSize = (WORD) (extraSize + extraDataBytes);
 		this->FormatTag = WaveFormat::Pcm;
 		this->Channels = 0;
 		this->SamplesPerSecond = 0;
